Add VotesController and specs

+++ db/migrate/20211012093720_create_votes.rb
@@ -1,7 +1,7 @@
 class CreateVotes < ActiveRecord::Migration[6.1]
   def change
     create_table :votes do |t|
-      t.decimal :rating
+      t.integer :rating
       t.references :user, null: false, foreign_key: true
       t.references :movie, null: false, foreign_key: true
